\name{guadeloupe}
\alias{guadeloupe}
\docType{package}
\title{
Centroids of buildings of Guadeloupe (French Island in Caribbean)
}

\description{
The guadeloupe centroids has been extracted from OpenStreetMap.

Map data copyrighted OpenStreetMap contributors and available from 'http://www.openstreetmap.org'


}

\value{
  \item{x}{longitude}
  \item{y}{latitude}

 }

 
\examples{
\dontrun{
data("guadeloupe")

plot(guadeloupe)
}
}
